fix: refactore search-result by making it server component, move logic from search/route directly into component
