+++ app/api/properties/search/route.js
@@ -1,43 +1,46 @@
-import connectDB from '@/config/database'
-import Property from '@/models/Property'
+// TO_REMOVE: not used in the app anymore
 
-export const dynamic = 'force-dynamic'
+// import connectDB from '@/config/database'
+// import Property from '@/models/Property'
+
+// export const dynamic = 'force-dynamic'
+
 // GET /api/properties/search
-export const GET = async (request) => {
-  try {
-    await connectDB()
+// export const GET = async (request) => {
+//   try {
+//     await connectDB()
 
-    const { searchParams } = new URL(request.url)
-    const location = searchParams.get('location')
-    const propertyType = searchParams.get('propertyType')
+//     const { searchParams } = new URL(request.url)
+//     const location = searchParams.get('location')
+//     const propertyType = searchParams.get('propertyType')
 
-    const locationPattern = new RegExp(location, 'i')
+//     const locationPattern = new RegExp(location, 'i')
 
-    // match location pattern against database fields
-    let query = {
-      $or: [
-        { name: locationPattern },
-        { description: locationPattern },
-        { 'location.street': locationPattern },
-        { 'location.city': locationPattern },
-        { 'location.state': locationPattern },
-        { 'location.zipcode': locationPattern },
-      ],
-    }
+//     // match location pattern against database fields
+//     let query = {
+//       $or: [
+//         { name: locationPattern },
+//         { description: locationPattern },
+//         { 'location.street': locationPattern },
+//         { 'location.city': locationPattern },
+//         { 'location.state': locationPattern },
+//         { 'location.zipcode': locationPattern },
+//       ],
+//     }
 
-    // only check for property if its not 'All'
-    if (propertyType && propertyType !== 'All') {
-      const typePattern = new RegExp(propertyType, 'i')
-      query.type = typePattern
-    }
+//     // only check for property if its not 'All'
+//     if (propertyType && propertyType !== 'All') {
+//       const typePattern = new RegExp(propertyType, 'i')
+//       query.type = typePattern
+//     }
 
-    const properties = await Property.find(query)
+//     const properties = await Property.find(query)
 
-    return new Response(JSON.stringify(properties), {
-      status: 200,
-    })
-  } catch (error) {
-    console.log(error)
-    return new Response('Something went wrong', { status: 500 })
-  }
-}
+//     return new Response(JSON.stringify(properties), {
+//       status: 200,
+//     })
+//   } catch (error) {
+//     console.log(error)
+//     return new Response('Something went wrong', { status: 500 })
+//   }
+// }
