+++ app/properties/search-results/page.jsx
@@ -1,41 +1,40 @@
-'use client'
-import { useState, useEffect } from 'react'
-import { useSearchParams } from 'next/navigation'
 import Link from 'next/link'
+
+import connectDB from '@/config/database'
+import Property from '@/models/Property'
+import { convertToSerializeableObject } from '@/utils/convertToObject'
+
 import { FaArrowAltCircleLeft } from 'react-icons/fa'
+
 import PropertyCard from '@/components/PropertyCard'
 import PropertySearchForm from '@/components/PropertySearchForm'
-import Spinner from '@/components/Spinner'
 
-const SeachResultsPage = () => {
-  const searchParams = useSearchParams()
-  const [properties, setProperties] = useState([])
-  const [loading, setLoading] = useState(true)
+const SearchResultsPage = async ({
+  searchParams: { location, propertyType },
+}) => {
+  await connectDB()
 
-  const location = searchParams.get('location')
-  const propertyType = searchParams.get('propertyType')
+  const locationPattern = new RegExp(location, 'i')
 
-  useEffect(() => {
-    const getSearchResults = async () => {
-      try {
-        const res = await fetch(
-          `/api/properties/search?location=${location}&propertyType=${propertyType}`
-        )
+  let query = {
+    $or: [
+      { name: locationPattern },
+      { description: locationPattern },
+      { 'location.street': locationPattern },
+      { 'location.city': locationPattern },
+      { 'location.state': locationPattern },
+      { 'location.zipcode': locationPattern },
+    ],
+  }
 
-        if (res.status === 200) {
-          const data = await res.json()
-          setProperties(data)
-        } else {
-          setProperties([])
-        }
-      } catch (error) {
-        console.log(error)
-      } finally {
-        setLoading(false)
-      }
-    }
-    getSearchResults()
-  }, [location, propertyType])
+  // check for property if its not 'All'
+  if (propertyType && propertyType !== 'All') {
+    const typePattern = new RegExp(propertyType, 'i')
+    query.type = typePattern
+  }
+
+  const propertiesQueryResults = await Property.find(query).lean()
+  const properties = convertToSerializeableObject(propertiesQueryResults)
 
   return (
     <>
@@ -44,32 +43,27 @@ const SeachResultsPage = () => {
           <PropertySearchForm />
         </div>
       </section>
-      {loading ? (
-        <Spinner loading={loading} />
-      ) : (
-        <section className='px-4 py-6'>
-          <div className='container-xl lg:container m-auto px-4 py-6'>
-            <Link
-              href='/properties'
-              className='flex items-center text-blue-500 hover:underline mb-3'
-            >
-              <FaArrowAltCircleLeft className='mr-2 mb-1' /> Back to properties
-            </Link>
-            <h1 className='text-2xl mb-4'>Search Results</h1>
-            {properties.length === 0 ? (
-              <p>No properties found</p>
-            ) : (
-              <div className='grid grid-cols-1 md:grid-cols-3 gap-6'>
-                {properties.map((property, index) => (
-                  <PropertyCard property={property} key={index} />
-                ))}
-              </div>
-            )}
-          </div>
-        </section>
-      )}
+      <section className='px-4 py-6'>
+        <div className='container-xl lg:container m-auto px-4 py-6'>
+          <Link
+            href='/properties'
+            className='flex items-center text-blue-500 hover:underline mb-3'
+          >
+            <FaArrowAltCircleLeft className='mr-2 mb-1' /> Back To Properties
+          </Link>
+          <h1 className='text-2xl mb-4'>Search Results</h1>
+          {properties.length === 0 ? (
+            <p>No search results found</p>
+          ) : (
+            <div className='grid grid-cols-1 md:grid-cols-3 gap-6'>
+              {properties.map((property) => (
+                <PropertyCard key={property._id} property={property} />
+              ))}
+            </div>
+          )}
+        </div>
+      </section>
     </>
   )
 }
-
-export default SeachResultsPage
+export default SearchResultsPage
